requirements.txt
# pip install trafilatura requests
google-generativeai,
trafilatura,
jinja2,
markdown,
requests,
smtplib,
email,
html,
css,
javascript,
json,
yaml,
xml,